+++ test/backdrop_test.dart
@@ -1,9 +1,8 @@
+import 'package:backdrop/backdrop.dart';
 import 'package:test/test.dart';
 
-import 'package:backdrop/backdrop.dart';
-
 void main() {
-  test('adds one to input values', () {
-
+  test('write tests here', () {
+    BackdropScaffold();
   });
 }